+++ src/ecallisto_ng/__init__.py
@@ -15,4 +15,4 @@ working agreement.
 
 from __future__ import annotations
 
-__version__ = "0.5.3"
+__version__ = "0.6.0"